Handle UnknownHostException in ExceptionHandlerImpl
This commit introduces specific handling for `UnknownHostException` in the `ExceptionHandlerImpl`. Now, if a network error is due to an inability to resolve the host, a more informative error message regarding internet connectivity will be displayed to the user in both Turkish and English.

The app's minor version has also been incremented to 53.